@@ -2,19 +2,21 @@
 using CerbiSharp.Tools.WebApiGenerator.Generator;
 using Microsoft.Extensions.Configuration;
 
-var paths = new List<string>();
 
-paths.Add(await CreateSampleSwaggerClientAsync());
+var paths = new List<string>
+{
+    await CreateSampleSwaggerClientAsync()
+};
 
 paths.ForEach(e => Console.WriteLine(e));
 
 
 static async Task<string> CreateSampleSwaggerClientAsync()
 {
-    const string NAME_SPACE = "CerbiSharp.Tools.WepApiGenerator.SampleSwagger";
+    const string NAME_SPACE = "CerbiSharp.Tools.WebApiGenerator.SampleSwagger";
     const string FILE_NAME = "SampleSwaggerClient.cs";
 
-    var swaggerJsonUrl = (GetAppSettings()?.GetSection("Urls:SampleSwagger").Get<string>()) ?? string.Empty;
+    var swaggerJsonUrl = (GetAppSettings()?.GetSection("URLs:SampleSwagger").Get<string>()) ?? string.Empty;
 
     return await new ApiClientGeneratorConfig()
         .SetDefaultApiClientGeneratorSettings()
@@ -27,19 +29,27 @@ static async Task<string> CreateSampleSwaggerClientAsync()
 
             settings.CSharpGeneratorSettings.Namespace = NAME_SPACE;
 
-            // add your configuration.
+            //      Add your configuration.
             settings.ClientClassAccessModifier = "internal";
             settings.ClientBaseClass = "SampleSwaggerBaseClient";
 
-            // An interface is used here for dependency injection.
+            //      An interface is used here for dependency injection.
             settings.ConfigurationClass = "SampleSwaggerConfig";
             settings.UseHttpClientCreationMethod = true;
 
-            // Setting name for response class.Default: SwaggerResponse.
-            // setting.ResponseClass = "SwaggerResponseLegacy";
+            //      Setting name for response class.Default: SwaggerResponse.
+            //setting.ResponseClass = "SwaggerResponseLegacy";
 
-            // Setting name generator.Default: ApiOperationNameGenerator.
-            // setting.CSharpGeneratorSettings.TypeNameGenerator = new LegacyTypeNameGenerator();
+            //      Setting name generator.Default: ApiOperationNameGenerator.
+            //setting.CSharpGeneratorSettings.TypeNameGenerator = new LegacyTypeNameGenerator();
+
+            //      Injecting Http Client
+            //settings.InjectHttpClient = true;
+
+            //      Removing config and Base class
+            settings.UseBaseUrl = false;
+            //settings.ClientBaseClass = string.Empty;
+            //settings.ConfigurationClass = string.Empty;
         })
         .GenerateClientFileAndSaveAsync();
 }
